[JENKINS-34150] Fixed issue by not checking the file
The issue itself seems to happen in case the check of the file presence happens
at the same moment as the file being created. This causes the batch file to not
create the file in question.

The solution to the issue is to track the alive state of the process that is
launched to execute the script. Only if that script is terminated the file is
checked.