[JENKINS-34150] Fix implementation using START
A fix for the issue was implemented using the start command. This has
the downside that the spawned processes create a CMD window for the
user that is used to execute the script.

Also a issue was fixed that caused a error while parsing the error code
in case the error file was written just in the same moment as the error
file was checked. In this cases the check did read a empty string from
the file and issued a error.

The check for the hanging issue was rewritten to cause the issue more
reliable by running multiple scripts at the same time.