Use waitpid to poll for process liveness
This reverts commit 3c10cc0cb0a4738646966494fbeaab54bca115e1 which
move the process liveness checking implementation from JNR to JNA.
This broke on Windows as there is no shared libc available.

Instead, we move back to JNR, but using waitpid instead of getpgid to
perform the polling. This is a generic enough interfacec that it should
have very consistent support across POSIX operating systems and can even
be supported on Windows (in a future JNR release).

To avoid class loading trouble with older versions of jnr-posix, we use
reflection to pick out waitpid support, which wasn't supported on Windows until
3.0.42.

Fixes JENKINS-38682.