always force an interpreter
Pipeline jobs including an sh step are including a shebang on the underlying script. Such shebang caused the script to be mode changed to executable. These "executables", when written on a mounted filesystem with the noexec flag, become non-executable (causing a not obvious failure).

Replacing the bare executable script with a call to the runner defined at the shebang, avoids this situation